# openapi.yaml
openapi: 3.1.0
info:
  title: Payment Service
  description: 'Production-ready Digital Wallet payment API. Provides a complete payment processing API including checkouts,
    refunds, customer management, and receipt generation.


    ## Authentication

    All API endpoints (except health checks and auth) require an `X-API-Key` header for authentication. Click the **Authorize**
    button (🔒) above to enter your API key, and it will be included in all requests.


    - **Admin key**: Set via `ADMIN_API_KEY` env var — used to manage tenant API keys

    - **Tenant keys**: Created via `POST /api/v1/admin/api-keys` — used for payment operations


    ## Idempotency

    POST endpoints support the `Idempotency-Key` header to prevent duplicate operations.


    ## Rate Limiting

    API requests are rate-limited per API key. Each tenant key can have custom rate limits. Check the `X-RateLimit-*` response
    headers for current limits.'
  version: 1.0.0
servers:
- url: http://localhost:8000
  description: Local development server
paths:
...
  /api/v1/checkout/sessions:
    post:
      tags:
      - Checkout
      summary: Create a custom checkout session
      description: 'Create a hosted checkout session. Returns a `checkout_url` — redirect the end-user''s browser to that
        URL to collect their custom Digital Wallet authorization.


        **This is the main integration point for client services.** The client only needs to call this endpoint, redirect
        the user to the local UI, and wait for the success callback redirect.'
      operationId: create_checkout_session_api_v1_checkout_sessions_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckoutSessionCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckoutSessionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/admin/api-keys:
    post:
      tags:
      - API Keys
      summary: Create a new API key
      description: 'Generate a new API key for a client/tenant. The raw key is returned **only in this response** — store
        it securely. It cannot be retrieved again.


        **Requires admin API key.**'
      operationId: create_api_key_api_v1_admin_api_keys_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiKeyCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyCreatedResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - API Keys
      summary: List API keys
      description: Retrieve a paginated list of all API keys. **Requires admin API key.**
      operationId: list_api_keys_api_v1_admin_api_keys_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: Items per page
          default: 20
          title: Limit
        description: Items per page
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          description: Items to skip
          default: 0
          title: Offset
        description: Items to skip
      - name: is_active
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          description: Filter by active status
          title: Is Active
        description: Filter by active status
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    ApiKeyCreate:
      properties:
        client_name:
          type: string
          maxLength: 255
          minLength: 1
          title: Client Name
          description: Name of the client/tenant
        description:
          anyOf:
          - type: string
            maxLength: 500
          - type: 'null'
          title: Description
          description: Description of what this key is used for
        scopes:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Scopes
          description: 'List of allowed scopes. If empty/null, all scopes are allowed. Available scopes: payments:read, payments:write,
            refunds:read, refunds:write, customers:read, customers:write'
        rate_limit_requests:
          anyOf:
          - type: integer
            maximum: 10000.0
            minimum: 1.0
          - type: 'null'
          title: Rate Limit Requests
          description: Custom rate limit (requests per window). Uses global default if not set.
        rate_limit_window_seconds:
          anyOf:
          - type: integer
            maximum: 3600.0
            minimum: 1.0
          - type: 'null'
          title: Rate Limit Window Seconds
          description: Custom rate limit window in seconds. Uses global default if not set.
        expires_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Expires At
          description: Expiration timestamp (UTC). Null means the key never expires.
      type: object
      required:
      - client_name
      title: ApiKeyCreate
      description: Request body to create a new API key.
      examples:
      - client_name: Ticket Service
        description: API key for the event ticketing microservice
        rate_limit_requests: 200
        rate_limit_window_seconds: 60
        scopes:
        - payments:read
        - payments:write
        - customers:read
    ApiKeyCreatedResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        key_prefix:
          type: string
          title: Key Prefix
          description: First characters of the key for identification
        client_name:
          type: string
          title: Client Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        scopes:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Scopes
        is_active:
          type: boolean
          title: Is Active
        rate_limit_requests:
          anyOf:
          - type: integer
          - type: 'null'
          title: Rate Limit Requests
        rate_limit_window_seconds:
          anyOf:
          - type: integer
          - type: 'null'
          title: Rate Limit Window Seconds
        last_used_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Used At
        expires_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Expires At
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        raw_key:
          type: string
          title: Raw Key
          description: The full API key. Store this securely — it cannot be retrieved again.
      type: object
      required:
      - id
      - key_prefix
      - client_name
      - is_active
      - created_at
      - updated_at
      - raw_key
      title: ApiKeyCreatedResponse
      description: 'Response returned only at creation time.

        Contains the raw API key — this is the ONLY time the full key is shown.'
    ApiKeyListResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ApiKeyResponse'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        limit:
          type: integer
          title: Limit
        offset:
          type: integer
          title: Offset
        has_more:
          type: boolean
          title: Has More
      type: object
      required:
      - items
      - total
      - limit
      - offset
      - has_more
      title: ApiKeyListResponse
      description: Paginated list of API keys.
    ApiKeyResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        key_prefix:
          type: string
          title: Key Prefix
          description: First characters of the key for identification
        client_name:
          type: string
          title: Client Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        scopes:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Scopes
        is_active:
          type: boolean
          title: Is Active
        rate_limit_requests:
          anyOf:
          - type: integer
          - type: 'null'
          title: Rate Limit Requests
        rate_limit_window_seconds:
          anyOf:
          - type: integer
          - type: 'null'
          title: Rate Limit Window Seconds
        last_used_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Used At
        expires_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Expires At
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - key_prefix
      - client_name
      - is_active
      - created_at
      - updated_at
      title: ApiKeyResponse
      description: API key details (without the raw key).
...
    CheckoutLineItem:
      properties:
        name:
          type: string
          maxLength: 200
          title: Name
          description: Product/item name shown on the checkout page
        quantity:
          type: integer
          minimum: 1.0
          title: Quantity
          description: Number of units
          default: 1
        price:
          type: integer
          exclusiveMinimum: 0.0
          title: Price
          description: Unit price in smallest currency unit (e.g., cents)
      type: object
      required:
      - name
      - price
      title: CheckoutLineItem
      description: A single line item in the checkout.
    CheckoutSessionCreate:
      properties:
        line_items:
          items:
            $ref: '#/components/schemas/CheckoutLineItem'
          type: array
          minItems: 1
          title: Line Items
          description: Items being purchased (displayed on the checkout page)
        currency:
          type: string
          maxLength: 3
          minLength: 3
          title: Currency
          description: ISO 4217 currency code
          default: usd
        success_url:
          type: string
          title: Success Url
          description: URL to redirect the customer after successful payment.
        cancel_url:
          type: string
          title: Cancel Url
          description: URL to redirect the customer if they cancel the checkout.
        customer_email:
          type: string
          title: Customer Email
          description: Email of the customer. If they don't exist in the Payment Service, a new identity will be created automatically.
        customer_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Customer Name
          description: Name of the customer (used if creating a new identity)
        metadata:
          anyOf:
          - type: object
          - type: 'null'
          title: Metadata
          description: Arbitrary key-value metadata passed to the payment
      type: object
      required:
      - line_items
      - success_url
      - cancel_url
      - customer_email
      title: CheckoutSessionCreate
      description: 'Request body to create a hosted checkout session.


        The client only needs to send the payment details and redirect URLs.

        The response includes a checkout_url — redirect the end-user there.'
      examples:
      - cancel_url: https://mytickets.com/order/123/cancel
        currency: usd
        customer_email: alice@example.com
        customer_name: Alice Smith
        line_items:
        - name: VIP Concert Ticket
          price: 5000
          quantity: 2
        metadata:
          event_id: evt_456
          order_id: order_123
        success_url: https://mytickets.com/order/123/success
    CheckoutSessionResponse:
      properties:
        session_id:
          type: string
          title: Session Id
          description: Checkout Session ID
        line_items:
          anyOf:
          - items:
              $ref: '#/components/schemas/CheckoutLineItem'
            type: array
          - type: 'null'
          title: Line Items
          description: Items to display in the UI
        checkout_url:
          type: string
          title: Checkout Url
          description: URL to redirect the customer to for payment
        status:
          type: string
          title: Status
          description: 'Session status: open, complete, or expired'
        success_url:
          type: string
          title: Success Url
          description: URL to redirect on success
        cancel_url:
          type: string
          title: Cancel Url
          description: URL to redirect on cancel
        expires_at:
          type: string
          format: date-time
          title: Expires At
          description: When this checkout session expires
        payment_status:
          type: string
          title: Payment Status
          description: 'Payment status: unpaid, paid, or no_payment_required'
        amount_total:
          type: integer
          title: Amount Total
          description: Total amount in smallest currency unit
        currency:
          type: string
          title: Currency
          description: ISO 4217 currency code
        customer_name:
          type: string
          title: Customer Name
          description: Customer name
        customer_email:
          type: string
          title: Customer Email
          description: Customer email
        metadata:
          anyOf:
          - type: object
          - type: 'null'
          title: Metadata
          description: Metadata
      type: object
      required:
      - session_id
      - checkout_url
      - status
      - success_url
      - cancel_url
      - expires_at
      - payment_status
      - amount_total
      - currency
      - customer_name
      - customer_email
      title: CheckoutSessionResponse
      description: 'Response after creating a checkout session.


        Redirect the end-user''s browser to `checkout_url` to complete payment.'
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Enter your API key. Use the **admin key** (from ADMIN_API_KEY env var) for `/api/v1/admin/*` endpoints,
        or a **tenant key** (ps_live_...) for payment/refund/customer operations.
...
security:
- ApiKeyAuth: []